+++ g33/CHANGELOG.md
@@ -4,6 +4,11 @@ This module adds custom G33 code to the µCNC parser. This similar to Linux CNC 
 
 ## Changelog
 
+### 2026-06-10 (#107) [@HuubBuis](https://github.com/HuubBuis)
+
+- improved threading synchronization speed (#107)
+- fixed threading synchronization status was not reset after a soft reset (#107)
+
 ### 2026-04-06
 
 - new math motion equations to produce an accurate pitch threading start synchronization